fixes #7 :  'items' return always the first key of a given type associated with a value
Internally changes the key-dicts by type to use OrderedDicts.
Therefore, when requesting the items(key_type=str) for a value defined as
   m["abc", "def"] = 123
The returned item will always have the key "abc" for 123, and
never "def". Previous behavior was undefined.

Also changes tests that relied on the return order of
"items": The multi_dict is not expected to work
as an OrderedDict itself, and code should not rely on this.